Backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreschema==0.0.4
Django==3.2.10
gunicorn==20.1.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.9
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
python-dotenv==0.21.1

fpdf2==2.5.6